save some bytes by replacing int with uint8_t in Prints, digitalRead() and analogWrite()

+++ hardware/arduino/avr/cores/arduino/wiring_digital.c
@@ -162,7 +162,7 @@ void digitalWrite(uint8_t pin, uint8_t val)
 	SREG = oldSREG;
 }
 
-int digitalRead(uint8_t pin)
+uint8_t digitalRead(uint8_t pin)
 {
 	uint8_t timer = digitalPinToTimer(pin);
 	uint8_t bit = digitalPinToBitMask(pin);
